fix(web): complete marketplace installs in a dialog
